refactor: SubCategoryMenu 컴포넌트 최적화 - expandedCategories를 Set<number>로 변경하여 toString() 변환 제거 - py-[15px] 중복 제거로 CSS 클래스 정리
