+++ src/components/common/examples/PaginationExample.tsx
@@ -3,7 +3,14 @@
 import { useState } from "react";
 import Pagination from "../Pagination";
 
-const PaginationExample = () => {
+/**
+ * Pagination 컴포넌트 사용 예시
+ * 
+ * @description
+ * 페이지네이션 컴포넌트의 기본적인 사용법을 보여주는 예시입니다.
+ */
+
+export default function PaginationExample() {
   const [currentPage, setCurrentPage] = useState(1);
   const totalPages = 10;
 
@@ -17,6 +24,4 @@ const PaginationExample = () => {
       <div>현재 페이지: {currentPage}</div>
     </div>
   );
-};
-
-export default PaginationExample; 
+}
